SQL - Data Cleaning: World Layoff
Script for data cleaning world layoff. This project focuses on cleaning and preparing world layoff data using SQL. The goal is to structure and standardise the data for better analysis and insights. Key tasks include removing duplicates, standardising formats, and handling null values using various SQL techniques like ROW_NUMBER, JOIN, and CASE statements. This hands-on experience is crucial for data analysts as it emphasises the importance of clean data for accurate analysis and reporting.