Project 1 - SQL code
The SQL code focuses on transforming and cleaning London job data to make it ready for analysis. Key steps include inspecting the dataset to understand its structure, creating a clean version of the table to preserve original data integrity, and applying common cleaning techniques such as removing duplicates, standardising formats, and handling missing values. This ensures the data is consistent and reliable for use in further visualisation, on tableau, and analysis stages.